Basic HTML Website Updated(Add more content to index.html)
Finished updating the basic HTML website by adding more content to the index.html file.
Removed the link to the external stylesheet as it is no longer needed.

What to do next:
Review the changes made to the index.html file to ensure everything is correct.
Change some of the divs to semantic HTML5 elements for better structure and accessibility.
Add the correct links for the navigation menu to point to the respective sections of the page.